+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ## Added
  * This CHANGELOG.md file
+ * API interchangeable between `SecureString` and `String`.
 
 ## [1.1.0](https://github.com/Password4j/password4j/releases/tag/1.1.0) - (2020-03-14)
 ### Added